requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
boto3
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7

